requirements.txt
dash
dash-bootstrap-components
pandas
mysql-connector-python
pymongo
neo4j